Fix completed button bug
The complete button updated on the active route and the completed route,
but not on the root route. I implemented a version of the root route on
a new route, /all, and used a store to try to make it reactive. That
still had the same effect. I then made the conditional statement that
determines the button color explicitly reactive and that fixed the bug.

I have not decided if I will reconfigure my root, active, and completed
routes to use the store or if I will leave them as is